Reorganized TutorialTriggers code
Made all the triggers and UI and animators into a list in my class. Where then each trigger needs to know which trigger it is(index in list). and then the triggers change the UI or make animations, have to know only the two UI needed, or again which animator to use(index).