Added Target script + upgrade platform
TargetTrigger script can be put on a trigger and can spawn as many targets as you want, and it updates the game manager about it. Then you can have a platform and set its needsTargets variable to true so that there has to be no more active targets before the platform moves.