Small change to checkpoint and Spider spawner
checkpoint model can now be null so you can just have the trigger as an invisible checkpoint. Spider spawner player in range variable now gets set to false when the player dies... Before if you died it would not register as an OnTriggerExit so it would keep spawning spiders and there would be too many to handle when you get back after respawning.